@@ -12,14 +12,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^17.0.6",
-    "@angular/compiler": "^17.0.6",
-    "@angular/core": "^17.0.6",
-    "@angular/forms": "^17.0.6",
-    "@angular/platform-browser": "^17.0.6",
-    "@angular/platform-browser-dynamic": "^17.0.6",
-    "@angular/router": "^17.0.6",
-    "@angular/service-worker": "17.0.6",
+    "@angular/common": "^18.2.0",
+    "@angular/compiler": "^18.2.0",
+    "@angular/core": "^18.2.0",
+    "@angular/forms": "^18.2.0",
+    "@angular/platform-browser": "^18.2.0",
+    "@angular/platform-browser-dynamic": "^18.2.0",
+    "@angular/router": "^18.2.0",
+    "@angular/service-worker": "18.2.0",
     "body-parser": "^1.19.0",
     "bootstrap": "^3.4.1",
     "core-js": "^2.6.9",
@@ -28,12 +28,12 @@
     "rxjs": "6.5.4",
     "tslib": "^2.0.0",
     "web-push": "^3.3.5",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.6",
-    "@angular/cli": "^17.0.6",
-    "@angular/compiler-cli": "^17.0.6",
+    "@angular-devkit/build-angular": "^18.1.4",
+    "@angular/cli": "^18.1.4",
+    "@angular/compiler-cli": "^18.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^6.14.6",
     "blue-harvest": "^0.2.1",
@@ -47,6 +47,6 @@
     "protractor": "^7.0.0",
     "ts-node": "5.0.1",
     "tslint": "~6.1.0",
-    "typescript": "5.2.2"
+    "typescript": "5.5.4"
   }
 }
