@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import {SwPush, SwUpdate} from '@angular/service-worker';
 import {interval} from 'rxjs';
-import {HttpClient} from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 
 @Component({
   selector: 'app-root',
